requirements/prod.txt
# Production dependencies
-r base.txt

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Security
pyjwt>=2.8.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Caching
redis>=5.0.0
hiredis>=2.2.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0
py-spy>=0.3.14

# Rate Limiting
slowapi>=0.1.9

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
aioboto3>=12.0.0
aiofiles>=23.2.0

# Cloud Providers
boto3>=1.33.0
azure-mgmt-compute>=30.4.0
azure-mgmt-resource>=23.0.0
azure-identity>=1.15.0
google-cloud-compute>=1.14.0
google-cloud-billing>=1.11.0
google-cloud-monitoring>=2.17.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Validation & Serialization
marshmallow>=3.20.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.0
tenacity>=8.2.0

# Background Tasks
celery>=5.3.0
flower>=2.0.0

# Testing (for production testing endpoints)
pytest-asyncio>=0.21.0

# Logging
python-json-logger>=2.0.7
sentry-sdk>=1.39.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
croniter>=2.0.0
humanize>=4.9.0